added filter tests, fixed some bugs
