+++ des/src/tracing/mod.rs
@@ -1,12 +1,12 @@
 //! Structured event tracing with custom context
 
+// mod filter;
 mod filter;
-mod filter2;
 mod format;
 mod output;
 mod policy;
 
-use self::{filter::TargetFilters, policy::DefaultScopeConfigurationPolicy};
+use self::{filter::Filters, policy::DefaultScopeConfigurationPolicy};
 use crate::{
     prelude::SimTime,
     sync::{Mutex, RwLock},
@@ -21,7 +21,10 @@ use std::{
 };
 use termcolor::BufferWriter;
 use tracing::{
-    level_filters::STATIC_MAX_LEVEL, metadata::LevelFilter, span, subscriber::SetGlobalDefaultError,
+    level_filters::STATIC_MAX_LEVEL,
+    metadata::LevelFilter,
+    span,
+    subscriber::{with_default, SetGlobalDefaultError},
 };
 
 pub use self::format::ColorfulTracingFormatter;
@@ -52,6 +55,8 @@ pub fn new_scope(s: &str) -> ScopeToken {
     let lock = SCOPES.lock();
     if let Some(scopes) = &*lock {
         scopes.send((token, s.to_string())).expect("Failed to send");
+    } else {
+        // WARNING MAYBE
     }
     token
 }
@@ -86,7 +91,7 @@ pub fn leave_scope() {
 pub struct Subscriber<P: ScopeConfigurationPolicy> {
     policy: P,
     scopes: RwLock<FxHashMap<u64, Scope>>,
-    filters: TargetFilters,
+    filters: Filters,
     max_log_level: LevelFilter,
 
     scopes_tx: Sender<(ScopeToken, String)>,
@@ -117,7 +122,7 @@ impl<P: ScopeConfigurationPolicy> Subscriber<P> {
         Self {
             policy,
             scopes: RwLock::new(FxHashMap::with_hasher(FxBuildHasher::default())),
-            filters: TargetFilters::new(true),
+            filters: Filters::from_env().expect("Failed to parse from env"),
             max_log_level: STATIC_MAX_LEVEL,
 
             scopes_tx,
@@ -140,7 +145,7 @@ impl<P: ScopeConfigurationPolicy> Subscriber<P> {
 
     /// Adds a target filter in textual repr to the subscriber.
     pub fn with_filter(mut self, filter: impl AsRef<str>) -> Self {
-        self.filters.parse_str(filter.as_ref());
+        self.filters = filter.as_ref().parse().expect("Failed to parse");
         self
     }
 
@@ -153,6 +158,18 @@ impl<P: ScopeConfigurationPolicy> Subscriber<P> {
         tracing::subscriber::set_global_default(self)?;
         let _ = SCOPES.lock().replace(tx);
         Ok(())
+    }
+
+    /// Set the subscriber as the default in the closure
+    pub fn with<R>(self, f: impl FnOnce() -> R) -> R
+    where
+        P: 'static,
+    {
+        let tx = self.scopes_tx.clone();
+        let prev = SCOPES.lock().replace(tx);
+        let ret = with_default(self, f);
+        *SCOPES.lock() = prev;
+        ret
     }
 }
 
@@ -212,13 +229,6 @@ impl<P: ScopeConfigurationPolicy + 'static> tracing::Subscriber for Subscriber<P
     fn event(&self, event: &tracing::Event<'_>) {
         // (-1) Target
 
-        let allowed_max_level = self
-            .filters
-            .filter_for(event.metadata().target(), LevelFilter::TRACE);
-        if allowed_max_level < *event.metadata().level() {
-            return;
-        }
-
         // (0) Identify current scope
         self.check_scopes();
         let scope_id = SCOPE_CURRENT_TOKEN.load(Ordering::SeqCst);
@@ -241,6 +251,14 @@ impl<P: ScopeConfigurationPolicy + 'static> tracing::Subscriber for Subscriber<P
             spans: &active,
             event,
         };
+
+        dbg!(&record);
+
+        let allowed_max_level = self.filters.level_filter_for(&record);
+        dbg!(&allowed_max_level);
+        if allowed_max_level < *event.metadata().level() {
+            return;
+        }
 
         if let Some(Scope { output, fmt, path }) = scope {
             // must be in this order because brwchk
